+++ react/javascript/src/components/gherkin/StatusIcon.tsx
@@ -8,7 +8,7 @@ import {
   IconDefinition,
 } from '@fortawesome/free-solid-svg-icons'
 import { messages } from '@cucumber/messages'
-import Status = messages.TestResult.Status
+import Status = messages.TestStepResult.Status
 import React from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import statusName from './statusName'